+++ firmware/NSPanelManagerFirmware/lib/NSPMConfig/NSPMConfig.h
@@ -103,7 +103,7 @@ public:
   bool relay2_default_mode;
 
   // @brief The temperature calibration in degrees
-  int temperature_calibration;
+  float temperature_calibration;
 
   // Indicates wether or not LittleFS was mounted or not.
   bool littlefs_mount_successfull;
